requirements.txt
flask
flask-cors
rembg
pillow